+++ lib/ui/orchestration.sh
@@ -76,6 +76,7 @@ install_tool() {
 	aria2) install_aria2 ;;
 	seleniumbase) install_seleniumbase ;;
 	playwright) install_playwright; _ensure_web_launchers ;;
+	browser_harness) install_browser_harness ;;
 	yandex_browser) install_yandex_browser; _ensure_web_launchers ;;
 	tor_browser) install_tor_browser ;;
 	qtox) install_qtox ;;